+++ routes/web.php
@@ -7,8 +7,12 @@ use App\Livewire\Admin\Agencies\Form as AgencyForm;
 use App\Livewire\Admin\Agencies\Show as AgencyShow;
 use App\Livewire\Admin\Agencies\Import as AgencyImport;
 use App\Livewire\Admin\DesignSystem;
+use App\Livewire\Admin\Users\Index as UsersIndex;
+use App\Livewire\Admin\Users\Form as UsersForm;
+use App\Livewire\Admin\Users\Show as UsersShow;
 use App\Livewire\PublicAgencies\Index as PublicAgenciesIndex;
 use App\Livewire\PublicAgencies\Show as PublicAgencyShow;
+use App\Livewire\Account\ChangePassword;
 use App\Modules\Agencies\Http\Controllers\AgencyImportPreviewController;
 use App\Modules\Agencies\Http\Controllers\AgencyMoveController;
 use App\Modules\Agencies\Models\Agency;
@@ -31,10 +35,15 @@ Route::get('/admin/agencies/import', AgencyImport::class)->middleware(['auth'])-
 Route::get('/admin/agencies/create', AgencyForm::class)->middleware(['auth'])->name('admin.agencies.create');
 Route::get('/admin/agencies/{agency}/edit', AgencyForm::class)->middleware(['auth'])->name('admin.agencies.edit');
 Route::get('/admin/agencies/{agency}', AgencyShow::class)->middleware(['auth'])->name('admin.agencies.show');
+Route::get('/admin/users', UsersIndex::class)->middleware(['auth'])->name('admin.users.index');
+Route::get('/admin/users/create', UsersForm::class)->middleware(['auth'])->name('admin.users.create');
+Route::get('/admin/users/{user}/edit', UsersForm::class)->middleware(['auth'])->name('admin.users.edit');
+Route::get('/admin/users/{user}', UsersShow::class)->middleware(['auth'])->name('admin.users.show');
 Route::get('/admin/design-system', DesignSystem::class)
     ->middleware(['auth'])
     ->name('admin.design-system');
 Route::get('/agencies', PublicAgenciesIndex::class)->name('public.agencies.index');
 Route::get('/agencies/{code}', PublicAgencyShow::class)->name('public.agencies.show');
+Route::get('/account/change-password', ChangePassword::class)->middleware(['auth'])->name('account.change-password');
 Route::post('/admin/agencies/import/preview', AgencyImportPreviewController::class)->middleware(['auth'])->name('admin.agencies.import.preview');
 Route::post('/admin/agencies/{agency}/move', AgencyMoveController::class)->middleware(['auth'])->name('admin.agencies.move');